requirements.txt
# Add your project dependencies here
fastapi
uvicorn
pytesseract
Pillow
transformers
torch
streamlit
requests
